Rename RemoteTabsFeature to SyncTabsFeature
This addresses issue #5272.
