+++ .buildconfig.yml
@@ -120,9 +120,9 @@ projects:
     path: components/feature/findinpage
     description: 'Feature that will subscribe to the selected session and show an UI for results of find in page.'
     publish: true
-  feature-remotetabs:
-    path: components/feature/remotetabs
-    description: 'Feature that provides access to other device''s tabs in the same account.'
+  feature-syncedtabs:
+    path: components/feature/syncedtabs
+    description: 'Feature that provides access to other devices'' tabs in the same account.'
     publish: true
   feature-sitepermissions:
     path: components/feature/sitepermissions
@@ -134,7 +134,7 @@ projects:
     publish: true
   feature-webcompat:
     path: components/feature/webcompat
-    description: 'Feature that provides hotfixes for websites from Mozillas Web Compatibility team'
+    description: 'Feature that provides hotfixes for websites from Mozilla''s Web Compatibility team'
     publish: true
   feature-addons:
     path: components/feature/addons
