+++ components/feature/syncedtabs/README.md
@@ -1,4 +1,4 @@
-# [Android Components](../../../README.md) > Feature > Remote Tabs
+# [Android Components](../../../README.md) > Feature > Synced Tabs
 
 Feature component for viewing tabs from other devices with a registered Fx Account.
 
@@ -9,23 +9,23 @@ Feature component for viewing tabs from other devices with a registered Fx Accou
 Use Gradle to download the library from [maven.mozilla.org](https://maven.mozilla.org/) ([Setup repository](../../../README.md#maven-repository)):
 
 ```Groovy
-implementation "org.mozilla.components:feature-remotetabs:{latest-version}"
+implementation "org.mozilla.components:feature-syncedtabs:{latest-version}"
 ```
 
 ## Usage
 
-In order to make use of the remote tab feature here, it's required to have an an FxA Account setup and Sync enabled.
+In order to make use of the synced tab feature here, it's required to have an an FxA Account setup and Sync enabled.
 See the [service-firefox-accounts](../../service/firefox-accounts/README.md) for more information how to set this up.
 
 ```kotlin
   // The feature will start listening to local tabs changes.
-  val remoteTabsFeature = RemoteTabsFeature(
+  val syncedTabsFeature = SyncedTabsFeature(
     accountManager = accountManager,
     store = browserStore,
     tabsStorage = tabsStorage
   )
   // Grab the list of opened tabs on other devices.
-  val otherDevicesTabs: Map<Device, List<Tab>> = remoteTabsFeature.getRemoteTabs()
+  val otherDevicesTabs: Map<Device, List<Tab>> = syncedTabsFeature.getSyncedTabs()
 
 ```
