+++ components/feature/syncedtabs/build.gradle
