+++ components/feature/syncedtabs/proguard-rules.pro
