+++ components/feature/syncedtabs/src/main/AndroidManifest.xml
