+++ components/feature/syncedtabs/src/test/java/mozilla/components/feature/syncedtabs/SyncedTabsFeatureTest.kt
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package mozilla.components.feature.remotetabs
+package mozilla.components.feature.syncedtabs
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import kotlinx.coroutines.ExperimentalCoroutinesApi
@@ -34,7 +34,7 @@ import org.mockito.Mockito.verify
 
 @ExperimentalCoroutinesApi
 @RunWith(AndroidJUnit4::class)
-class RemoteTabsFeatureTest {
+class SyncedTabsFeatureTest {
     private lateinit var store: BrowserStore
     private lateinit var tabsStorage: RemoteTabsStorage
     private lateinit var accountManager: FxaAccountManager
@@ -55,7 +55,7 @@ class RemoteTabsFeatureTest {
 
     @Test
     fun `listens to browser store changes and stores its state`() = runBlocking {
-        val feature = RemoteTabsFeature(accountManager, store, tabsStorage)
+        val feature = SyncedTabsFeature(accountManager, store, tabsStorage)
         feature.start()
         // This action won't change the state, but will run the flow.
         store.dispatch(TabListAction.RemoveAllPrivateTabsAction)
@@ -68,7 +68,7 @@ class RemoteTabsFeatureTest {
 
     @Test
     fun `stops listening to browser store changes on stop()`() = runBlocking {
-        val feature = RemoteTabsFeature(accountManager, store, tabsStorage)
+        val feature = SyncedTabsFeature(accountManager, store, tabsStorage)
         feature.start()
         // Run the flow.
         store.dispatch(TabListAction.RemoveAllPrivateTabsAction)
@@ -84,8 +84,8 @@ class RemoteTabsFeatureTest {
     }
 
     @Test
-    fun `getRemoteTabs matches tabs with FxA devices`() = runBlocking {
-        val feature = spy(RemoteTabsFeature(accountManager, store, tabsStorage))
+    fun `getSyncedTabs matches tabs with FxA devices`() = runBlocking {
+        val feature = spy(SyncedTabsFeature(accountManager, store, tabsStorage))
         val device1 = Device(
             id = "client1",
             displayName = "Foo Client",
@@ -117,19 +117,19 @@ class RemoteTabsFeatureTest {
         assertEquals(mapOf(
             device1 to tabsClient1,
             device2 to tabsClient2
-        ), feature.getRemoteTabs())
+        ), feature.getSyncedTabs())
     }
 
     @Test
-    fun `getRemoteTabs returns empty list if syncClients() is null`() = runBlocking {
-        val feature = spy(RemoteTabsFeature(accountManager, store, tabsStorage))
+    fun `getSyncedTabs returns empty list if syncClients() is null`() = runBlocking {
+        val feature = spy(SyncedTabsFeature(accountManager, store, tabsStorage))
         doReturn(null).`when`(feature).syncClients()
-        assertEquals(emptyMap<Device, List<Tab>>(), feature.getRemoteTabs())
+        assertEquals(emptyMap<Device, List<Tab>>(), feature.getSyncedTabs())
     }
 
     @Test
     fun `syncClients returns clients if the account is set and constellation state is set too`() {
-        val feature = spy(RemoteTabsFeature(accountManager, store, tabsStorage))
+        val feature = spy(SyncedTabsFeature(accountManager, store, tabsStorage))
         val account: OAuthAccount = mock()
         val constellation: DeviceConstellation = mock()
         val state: ConstellationState = mock()
@@ -152,7 +152,7 @@ class RemoteTabsFeatureTest {
 
     @Test
     fun `syncClients returns null if the account is set but constellation state is null`() {
-        val feature = spy(RemoteTabsFeature(accountManager, store, tabsStorage))
+        val feature = spy(SyncedTabsFeature(accountManager, store, tabsStorage))
         val account: OAuthAccount = mock()
         val constellation: DeviceConstellation = mock()
         whenever(accountManager.authenticatedAccount()).thenReturn(account)
@@ -163,7 +163,7 @@ class RemoteTabsFeatureTest {
 
     @Test
     fun `syncClients returns null if the account is null`() {
-        val feature = spy(RemoteTabsFeature(accountManager, store, tabsStorage))
+        val feature = spy(SyncedTabsFeature(accountManager, store, tabsStorage))
         whenever(accountManager.authenticatedAccount()).thenReturn(null)
         assertEquals(null, feature.syncClients())
     }
