+++ components/feature/syncedtabs/src/test/resources/mockito-extensions/org.mockito.plugins.MockMaker
